+++ CHANGELOG.md
@@ -1,6 +1,8 @@
 # Changelog
 
 ## [Unreleased]
+
+## [8.1.0]
 ### Added
 - [#79] Added option to hide progress bar, Thanks to [@techi602]
 
@@ -269,7 +271,8 @@ patchesJson6902:
 [#14]: https://github.com/sspooky13/yaml-standards/issues/14
 [#13]: https://github.com/sspooky13/yaml-standards/pull/13
 
-[Unreleased]: https://github.com/sspooky13/yaml-standards/compare/8.0.1...HEAD
+[Unreleased]: https://github.com/sspooky13/yaml-standards/compare/8.1.0...HEAD
+[8.1.0]: https://github.com/sspooky13/yaml-standards/compare/8.0.1...8.1.0
 [8.0.1]: https://github.com/sspooky13/yaml-standards/compare/8.0.0...8.0.1
 [8.0.0]: https://github.com/sspooky13/yaml-standards/compare/7.0.1...8.0.0
 [7.0.1]: https://github.com/sspooky13/yaml-standards/compare/7.0.0...7.0.1